- Organize Ruby primitive wrappers by file - Add small convenience APIs for primitives
